Print all 64 bits of seek position.
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@30662 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_lavf.c
@@ -96,7 +96,7 @@ static int mp_read(void *opaque, uint8_t *buf, int size) {
 static int64_t mp_seek(void *opaque, int64_t pos, int whence) {
     stream_t *stream = opaque;
     int64_t current_pos;
-    mp_msg(MSGT_HEADER,MSGL_DBG2,"mp_seek(%p, %d, %d)\n", stream, (int)pos, whence);
+    mp_msg(MSGT_HEADER,MSGL_DBG2,"mp_seek(%p, %"PRId64", %d)\n", stream, pos, whence);
     if(whence == SEEK_CUR)
         pos +=stream_tell(stream);
     else if(whence == SEEK_END && stream->end_pos > 0)
